Moved the shared functions to job_launcher_utils. Removed the improper inheritence.
