+++ nvflare/app_common/launchers/subprocess_launcher.py
@@ -24,7 +24,7 @@ from nvflare.apis.fl_context import FLContext
 from nvflare.apis.shareable import Shareable
 from nvflare.apis.signal import Signal
 from nvflare.app_common.abstract.launcher import Launcher, LauncherRunStatus
-from nvflare.private.fed.utils.fed_utils import add_custom_dir_to_path
+from nvflare.utils.job_launcher_utils import add_custom_dir_to_path
 
 
 def log_subprocess_output(process, logger):
